Add traffic and storage price query generation functions

+++ src/sc_keeper/parameters.py
@@ -401,7 +401,7 @@ storage_type = Annotated[
 ]
 
 extra_storage_size = Annotated[
-    Optional[float],
+    Optional[int],
     Query(
         title="Extra Storage Size",
         description=(
